Completed reverseString exercise.
I took 2 days to do it wrong. I should have turned
the string into an array, reversed the array and then
joined it back into a string.

Instead I've created a loop that for the length of the
input string would create a new array without the last
character while joining that character to the reverseString.

Unreasonably complicated way to get the job done.
It does get the job done though so I guess it's better
than nothing.